Updated and added public frameworks from Omni's current svn trunk.
Highlights include the following...

New code:
  - Added OmniUnzip, an ObjC wrapper for reading and writing zip files, based on 'minizip'.
  - Added OmniFileStore, which provides a NSFileManager-like interface, but using NSURLs, to find/read/write files across the local filesystem or WebDAV.

Configurations and Scripts:
  - Require 10.6
  - Use @rpath instead of @executable_path
  - Build with gcc 4.2 and the clang static analyzer (top-of-tree is much improved and we build with that too if you want to edit the configs)
  - Enabled 64->32 implicit conversion warnings. This catches lots of crufty code at the expense of some spurious warnings.
  - Added our workaround for an Xcode misfeature where unit test don't get run for release builds.

General:
  - Lots of 64-bit cleanup.
    - In some cases, we've just #if'd out code and replaced it with OBFinishPorting (which raises when we hit it).
    - Much of OWF and OIF, for example, are slated to be trimmed out since we no longer use them.
  - Removed lots of pre-10.6 code and switched to 10.6 variants where applicable. More left to be done.
  - Removed other code that we weren't using any more (instead of porting it w/o a real world test).
  - Various leak and bug fixes found via top-of-tree clang
  - Some work towards GC support.

OmniBase:
  - Extended and improved our runtime signature verification. This does things like ensure you didn't impelement a protocol and return a 32-bit type when you should have returned a 64-bit type, etc.
  - Factored C core API out of some code to better allow inclusion of only C files in system bundles (to avoid possibility of namespace conflicts)

OmniFoundation:
  - Factored C core API out of some code to better allow inclusion of only C files in system bundles (to avoid possibility of namespace conflicts)
  - Added OFExtent; NSRange but with floats.
  - OFHeap uses blocks now.
  - OFRandom is now a wrapper around "SIMD-oriented Fast Mersenne Twister" from <http://www.math.sci.hiroshima-u.ac.jp/~m-mat/MT/SFMT/index.html>
  - A bunch more work on signed XML support, CDSA wrappers.
  - Add some simple block category methods to containers. More obviously could be done, but we're only adding them as we use them.

OmniFileStore
  - First public release.
  - Vaguely NSFileManager-like API that handles local filesystems and WebDAV.
  - Includes support for https credentials, typically stored in Keychain, via a delegate (since the actual Keychain APIs differ between iPhone and Mac OS X).
  - Support for asynchronous operations (not all implemented).
  - Doesn't do everything, notably, no support for condiional GETs, properties, ETags.
  - Includes a 'ofs' command line tool, mostly for testing.

OmniUnzip:
  - Our ObjC wrapper around the 'minizip' tool.
  - Handles NSFileWrapper <-> zip archive
  - Handles regular, link and directory file types along with basic metadata.
  - Support for moving raw zip members between archives w/o decompressing and recompressing them.

OmniDataObjects:
  - Added support for declaring properties as being calculated. This makes the properties readonly.
  - For one-to-one relationships, require that exactly one side is marked calculated.
  - Added starts-with and contains predicate support for strings.
  - Added support for marking objects as undeletable.
  - Updated delete propagation to issue KVO for to-one relationships of the objects in the deletion set. Also, deleted objects return nil for to-one accessors instead of asserting and crashing.  This allows "foo.bar" key path observations to work (which we didn't use in previous versions).
  - Added some (partial) sanity checking that a predicate used for in-memory vs. SQL qualification will result in the same set of objets matching.
  - Unconditionally support undo, since the iPhone does now.
  - Removed our NSPredicate implementation, now that the iPhone has it (we still have support for building SQL qualifiers, of course).

OmniAppKit:
  - Added our (semi-experimental) OATargetSelection categories for customizing the responder chain and finding target for actions more robustly. YMMV.  See <http://www.cocoatoa.com/posts/2009/12/08/modernizing_the_responder_chain/> for some notes.
  - Some support for block-based NSAlert completion hooks.
  - More support for quarantine property operations via NSFileManager category methods.
  - NOTE: Many of the tests of the bezier extensions are currently failing due to precision changes when building 64-bit. More updates in the future...

OmniSoftwareUpdate:
  - More work on signed feeds
  - Support for compressing common OpenGL extensions into a bitset.
  - Fixes to querying display properties.
  - OpenCL details added to query
  - UI improvements
  - More fixes and bullet-proofing of install process
  - Beginnings of tool to automate appcast feed generation, including signature generation.

+++ Frameworks/OIF/OIAnimationInstance.h
@@ -1,4 +1,4 @@
-// Copyright 1997-2005 Omni Development, Inc.  All rights reserved.
+// Copyright 1997-2005, 2010 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -22,7 +22,7 @@
     int loopSeconds;                    // Total duration the animation should last before stopping
     unsigned int loopCount;             // Maximum number of loops to display
     unsigned int remainingLoops;        // Remaining number of loops to display, initialized from loopCount
-    unsigned int nextFrame;             // Index of animation frame to display
+    NSUInteger nextFrame;             // Index of animation frame to display
     OFScheduledEvent *nextFrameEvent;
     NSLock *nextFrameEventLock;
     NSTimer *expirationTimer;
