+++ Frameworks/OWF/Content.subproj/OWDataStreamFilterCursor.h
@@ -1,4 +1,4 @@
-// Copyright 2003-2005 Omni Development, Inc.  All rights reserved.
+// Copyright 2003-2005, 2010 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -14,7 +14,7 @@
 @interface OWDataStreamFilterCursor : OWDataStreamCursor
 {
     NSMutableData *bufferedData;
-    unsigned int bufferedDataStart, bufferedDataValidLength;
+    NSUInteger bufferedDataStart, bufferedDataValidLength;
     BOOL canFillMoreBuffer, haveStartedFilter;
 }
