+++ Frameworks/OWF/Content.subproj/OWDataStreamScanner.h
@@ -1,4 +1,4 @@
-// Copyright 1997-2005 Omni Development, Inc.  All rights reserved.
+// Copyright 1997-2005, 2010 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -16,9 +16,9 @@
     OWDataStreamCharacterCursor *streamCursor;
     
     unichar *buffer;
-    unsigned int bufferLength, bufferSize;	/* buffer length and size, in unichars */
-    unsigned int bufferOffset;			/* buffer start offset, in characters */
-    unsigned int minimumReadBufferLength;       /* how many characters to read at a time */
+    NSUInteger bufferLength, bufferSize;	/* buffer length and size, in unichars */
+    NSUInteger bufferOffset;			/* buffer start offset, in characters */
+    NSUInteger minimumReadBufferLength;       /* how many characters to read at a time */
 }
 
 - initWithCursor:(OWDataStreamCharacterCursor *)aStreamCursor bufferLength:(unsigned int)aBufferLength;
