+++ Frameworks/OWF/Content.subproj/OWObjectStreamCursor.h
@@ -1,4 +1,4 @@
-// Copyright 1997-2005 Omni Development, Inc.  All rights reserved.
+// Copyright 1997-2005, 2010 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -15,12 +15,12 @@
 {
     OWAbstractObjectStream *objectStream;
     void *hint;
-    unsigned int streamIndex;
+    NSUInteger streamIndex;
 }
 
 - initForObjectStream:(OWAbstractObjectStream *)anObjectStream;
 - (OWAbstractObjectStream *)objectStream;
-- (unsigned int)streamIndex;
+- (NSUInteger)streamIndex;
 
 - (id)readObject;
 - (void)skipObjects:(int)count;
