+++ Frameworks/OWF/Content.subproj/OWObjectStreamCursor.m
@@ -1,4 +1,4 @@
-// Copyright 1997-2005 Omni Development, Inc.  All rights reserved.
+// Copyright 1997-2005, 2010 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -52,7 +52,7 @@ RCS_ID("$Id$")
     return objectStream;
 }
 
-- (unsigned int)streamIndex;
+- (NSUInteger)streamIndex;
 {
     return streamIndex;
 }
@@ -81,7 +81,7 @@ RCS_ID("$Id$")
 
 // OWCursor subclass
 
-- (unsigned int)seekToOffset:(int)offset fromPosition:(OWCursorSeekPosition)position;
+- (NSUInteger)seekToOffset:(NSInteger)offset fromPosition:(OWCursorSeekPosition)position;
 {
     if (abortException)
 	[abortException raise];
