+++ Frameworks/OmniFoundation/Tests/OFXMLCursorTests.m
@@ -1,4 +1,4 @@
-// Copyright 2003-2008 Omni Development, Inc.  All rights reserved.
+// Copyright 2003-2008, 2010 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -75,7 +75,7 @@ RCS_ID("$Id$");
 - (void) testEnumerateChildren;
 {
     id child;
-    unsigned int childIndex, childCount;
+    NSUInteger childIndex, childCount;
     
     OFXMLCursor *cursor = [doc cursor];
